refactor(transactions): extract QuickTypeForm so create and edit share it
