+++ features/transactions/QuickEntry.tsx
@@ -1,124 +1,22 @@
 'use client';
 
 import { ChevronDownIcon, PlusIcon } from 'lucide-react';
-import { useState, useTransition } from 'react';
+import { useState } from 'react';
+import { QuickTypeForm } from './QuickTypeForm';
 import { createTransactionAction } from './actions';
 import { serializeDraftJson } from './entry/draftReducer';
-import { Field } from './entry/typeForms/fields';
-import type { HeaderFields } from './entry/types/adapter';
-import { QUICK_ENTRY_SPECS, type QuickEntrySpec } from './quickEntrySpecs';
+import { QUICK_ENTRY_SPECS } from './quickEntrySpecs';
 import { Button } from '@/components/ui/button';
-import {
-  Dialog,
-  DialogContent,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle,
-} from '@/components/ui/dialog';
+import { Dialog } from '@/components/ui/dialog';
 import {
   DropdownMenu,
   DropdownMenuContent,
   DropdownMenuItem,
   DropdownMenuTrigger,
 } from '@/components/ui/dropdown-menu';
-import { Input } from '@/components/ui/input';
-import { Label } from '@/components/ui/label';
 import { useRouter } from 'next/navigation';
 
 type Props = { accounts: string[]; defaultCurrency: string };
-
-/**
- * The dialog body for one entry type. Owns the field state and the save path
- * (compile via the spec's adapter → same server action as the full entry form),
- * so ledger stays the authority on whether the entry balances and is valid.
- */
-function QuickEntryContent({
-  spec,
-  accounts,
-  defaultCurrency,
-  onDone,
-}: Props & { spec: QuickEntrySpec<HeaderFields>; onDone: () => void }) {
-  const router = useRouter();
-  const [fields, setFields] = useState<HeaderFields>(() =>
-    spec.makeEmpty({ accounts, defaultCurrency })
-  );
-  const [pending, startTransition] = useTransition();
-  const [error, setError] = useState<string>();
-
-  const update = (patch: Partial<HeaderFields>) =>
-    setFields((f) => ({ ...f, ...patch }));
-
-  const save = () => {
-    const invalid = spec.validate(fields);
-    if (invalid) return setError(invalid);
-    startTransition(async () => {
-      const payee =
-        fields.payee.trim() || spec.resolvePayee?.(fields) || spec.label;
-      const draft = spec.compile({ ...fields, payee }, { defaultCurrency });
-      const formData = new FormData();
-      formData.set('draft', serializeDraftJson(draft, 'create'));
-      const result = await createTransactionAction(null, formData);
-      if (result.ok) {
-        onDone();
-        router.refresh();
-      } else {
-        const fieldError = result.fieldErrors
-          ? Object.values(result.fieldErrors).flat()[0]
-          : undefined;
-        setError(result.formError ?? fieldError ?? 'Could not save.');
-      }
-    });
-  };
-
-  return (
-    <DialogContent>
-      <DialogHeader>
-        <DialogTitle>
-          {spec.icon} {spec.label}
-        </DialogTitle>
-      </DialogHeader>
-
-      <div className="flex flex-col gap-4">
-        <spec.Fields
-          fields={fields}
-          update={update}
-          accounts={accounts}
-          defaultCurrency={defaultCurrency}
-        />
-
-        <div className="grid grid-cols-2 gap-3">
-          <Field label="Date">
-            <Input
-              type="date"
-              value={fields.date}
-              onChange={(e) => update({ date: e.target.value })}
-            />
-          </Field>
-          <div className="flex flex-col gap-1.5">
-            <Label htmlFor="quick-entry-payee">
-              Description{' '}
-              <span className="text-muted-foreground">(optional)</span>
-            </Label>
-            <Input
-              id="quick-entry-payee"
-              value={fields.payee}
-              onChange={(e) => update({ payee: e.target.value })}
-              placeholder="What was it for?"
-            />
-          </div>
-        </div>
-
-        {error && <p className="text-sm text-destructive">{error}</p>}
-      </div>
-
-      <DialogFooter showCloseButton>
-        <Button onClick={save} disabled={pending}>
-          {pending ? 'Saving…' : 'Save'}
-        </Button>
-      </DialogFooter>
-    </DialogContent>
-  );
-}
 
 /**
  * Split button in the app header: the primary click logs an expense; the caret
@@ -126,9 +24,18 @@ function QuickEntryContent({
  * from every page.
  */
 export default function QuickEntry({ accounts, defaultCurrency }: Props) {
+  const router = useRouter();
   const [active, setActive] = useState<string | null>(null);
   const [primary, ...rest] = QUICK_ENTRY_SPECS;
-  const spec = QUICK_ENTRY_SPECS.find((s) => s.kind === active) ?? null;
+  const spec = QUICK_ENTRY_SPECS.find((entry) => entry.kind === active) ?? null;
+
+  const onSave = async (draft: Parameters<typeof serializeDraftJson>[0]) => {
+    const formData = new FormData();
+    formData.set('draft', serializeDraftJson(draft, 'create'));
+    const result = await createTransactionAction(null, formData);
+    if (result.ok) router.refresh();
+    return result;
+  };
 
   return (
     <>
@@ -156,10 +63,13 @@ export default function QuickEntry({ accounts, defaultCurrency }: Props) {
             }
           />
           <DropdownMenuContent align="end">
-            {rest.map((s) => (
-              <DropdownMenuItem key={s.kind} onClick={() => setActive(s.kind)}>
-                <span className="mr-1">{s.icon}</span>
-                {s.label}
+            {rest.map((entry) => (
+              <DropdownMenuItem
+                key={entry.kind}
+                onClick={() => setActive(entry.kind)}
+              >
+                <span className="mr-1">{entry.icon}</span>
+                {entry.label}
               </DropdownMenuItem>
             ))}
           </DropdownMenuContent>
@@ -173,11 +83,12 @@ export default function QuickEntry({ accounts, defaultCurrency }: Props) {
         }}
       >
         {spec && (
-          <QuickEntryContent
+          <QuickTypeForm
             key={spec.kind}
             spec={spec}
             accounts={accounts}
             defaultCurrency={defaultCurrency}
+            onSave={onSave}
             onDone={() => setActive(null)}
           />
         )}
